+++ ts/features/bonus/bpd/saga/networking/loadPeriodsWithInfo.ts
@@ -1,16 +1,9 @@
 import { findFirst } from "fp-ts/lib/Array";
 import { Either, isRight, right } from "fp-ts/lib/Either";
-import { select, take } from "redux-saga-test-plan/matchers";
 import { all, call, put } from "redux-saga/effects";
-import { ActionType, getType } from "typesafe-actions";
 import { SagaCallReturnType } from "../../../../../types/utils";
-import { loadAbi } from "../../../../wallet/onboarding/bancomat/store/actions";
-import { abiSelector } from "../../../../wallet/onboarding/store/abi";
 import { BackendBpdClient } from "../../api/backendBpdClient";
-import { isReady } from "../../model/RemoteValue";
-import { bpdLoadActivationStatus } from "../../store/actions/details";
 import { bpdPeriodsAmountLoad } from "../../store/actions/periods";
-import { bpdTransactionsLoad } from "../../store/actions/transactions";
 import {
   BpdPeriodWithInfo,
   BpdRanking,
@@ -19,44 +12,6 @@ import {
 import { BpdAmount, BpdAmountError, bpdLoadAmountSaga } from "./amount";
 import { bpdLoadPeriodsSaga } from "./periods";
 import { bpdLoadRaking } from "./ranking";
-
-/**
- * Prefetch all the BPD details data:
- * - Activation Status
- * - Abi list
- * - Periods
- * - Amount foreach period !== "Inactive"
- * - Transactions foreach period !== "Inactive"
- */
-export function* loadBpdData() {
-  yield put(bpdLoadActivationStatus.request());
-
-  const abiList: ReturnType<typeof abiSelector> = yield select(abiSelector);
-
-  // The volatility of the bank list is extremely low.
-  // There is no need to refresh it every time.
-  // A further refinement could be to insert an expiring cache
-  if (!isReady(abiList)) {
-    yield put(loadAbi.request());
-  }
-
-  yield put(bpdPeriodsAmountLoad.request());
-
-  const periods: ActionType<
-    typeof bpdPeriodsAmountLoad.success | typeof bpdPeriodsAmountLoad.failure
-  > = yield take([
-    getType(bpdPeriodsAmountLoad.success),
-    getType(bpdPeriodsAmountLoad.failure)
-  ]);
-
-  if (periods.type === getType(bpdPeriodsAmountLoad.success)) {
-    yield all(
-      periods.payload
-        .filter(p => p.status !== "Inactive")
-        .map(period => put(bpdTransactionsLoad.request(period.awardPeriodId)))
-    );
-  }
-}
 
 /**
  * Load the periods list with the amount and ranking information foreach period (active or closed)
